app/requirements.txt
PyYAML==3.12
requests==2.19.1
marshmallow==2.15.3
slackclient==1.2.1
pytest==3.6.3